requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Providers
google-generativeai==0.3.1  # Gemini API
anthropic==0.7.8  # Claude API

# LiveKit (for real-time interviews)
livekit==0.10.0
livekit-agents==0.7.0
# livekit-plugins-hedra==0.1.0  # Hedra plugin (check Hedra docs for latest version)
livekit-plugins-openai==0.6.0
livekit-plugins-deepgram==0.5.0
livekit-plugins-elevenlabs==0.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Audio Processing (if using local Whisper)
openai-whisper==20231117  # Optional, can use API instead
torch==2.1.0  # For Whisper (CPU version)

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Optional: Database (for production)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # PostgreSQL
# alembic==1.12.1  # Database migrations

